Updated README.md for troubleshooting expected paths for currentFrame.txt and prompt.txt
I don't use the C: drive for my stable diffusion webui install and the C: drive is hardcoded for the paths for currentFrame.txt and prompt.txt in animation.py, render.py and main.py. I've included a quick troubleshooting step that will fix the "ALREADY LOCKED" spam and crash issue that occurs when the C: drive is not your local webui install path.